Update sshget
I am not sure how any of this really works, also not sure of the implications of the choices . main idea was getting quotes around paths to deal with spaces in file names. After that, around multiple paths. I guess i use your sshget script enough it was the least I could do give you the code for use as an idea starter, these probably are not the best way to deal with the goal